Removed Duplicacy and Made changes in TestDeleteClusterShouldResetSDKConnection
