+++ drivers/storage/portworx/util/util.go
@@ -785,6 +785,34 @@ func ParsePxProxyURL(proxy string) (string, string, string, error) {
 func getSpecsBaseDir() string {
 	return PortworxSpecsDir
 }
+func GetStorageNodes(
+	cluster *corev1.StorageCluster, k8sClient client.Client, sdkConn *grpc.ClientConn) ([]*api.StorageNode, error) {
+	sdkConn, err := GetPortworxConn(sdkConn, k8sClient, cluster.Namespace)
+	if err != nil {
+		if IsFreshInstall(cluster) && strings.HasPrefix(err.Error(), ErrMsgGrpcConnection) {
+			// Don't return grpc connection error during initialization,
+			// as SDK server won't be up anyway
+			logrus.Warn(err)
+			return []*api.StorageNode{}, nil
+		}
+		return nil, err
+	}
+
+	nodeClient := api.NewOpenStorageNodeClient(sdkConn)
+	ctx, err := SetupContextWithToken(context.Background(), cluster, k8sClient)
+	if err != nil {
+		return nil, err
+	}
+	nodeEnumerateResponse, err := nodeClient.EnumerateWithFilters(
+		ctx,
+		&api.SdkNodeEnumerateWithFiltersRequest{},
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return nodeEnumerateResponse.Nodes, nil
+}
 
 // GetPortworxConn returns a new Portworx SDK client
 func GetPortworxConn(sdkConn *grpc.ClientConn, k8sClient client.Client, namespace string) (*grpc.ClientConn, error) {
